Enhance subject consent form and admin functionality
This commit refines subject consent data handling in the admin and form views. It corrects data pull logic, allowing subject identifier to be fetched from the screening identifier when it's not directly available. Minor code redundancies were also removed, optimizing the overall codebase.
