+++ flourish_dashboard/utils.py
@@ -1,6 +1,8 @@
 from django.apps import apps as django_apps
 from edc_base.utils import age, get_utcnow
 
+caregiver_config = django_apps.get_app_config('flourish_caregiver')
+
 
 class FlourishDashboardUtils:
 
@@ -66,6 +68,22 @@ class FlourishDashboardUtils:
         else:
             return screening_obj.screeningpregwomeninline_set.filter(
                 child_subject_identifier=child_subject_identifier)
+
+    def consent_version_obj(self, screening_identifier=None):
+        consent_version_cls = django_apps.get_model(
+            'flourish_caregiver.flourishconsentversion')
+        try:
+            consent_version_obj = consent_version_cls.objects.get(
+                screening_identifier=screening_identifier)
+        except consent_version_cls.DoesNotExist:
+            return None
+        else:
+            return consent_version_obj
+
+    def is_latest_consent_version(self, screening_identifier=None):
+        consent_version_obj = self.consent_version_obj(screening_identifier)
+        return str(consent_version_obj.child_version) == str(
+            caregiver_config.consent_version)
 
 
 flourish_dashboard_utils = FlourishDashboardUtils()
